finished

+++ lib/main/data/models/product.dart
@@ -46,7 +46,7 @@ class Product extends Equatable {
         productMainUses: json['productMainUses'],
         productWorkCharacteristics: json['productWorkCharacteristics'],
         productId: json['productId'],
-        ratingNumbers: json['ratingNumbers'],
+    ratingNumbers: json['ratingNumbers'].toInt(),
         productRating: json['productRating'].toDouble(),
         productCity: json['productCity'],
         productNewPrice: json['productNewPrice'].toDouble(),
